requirements.txt
Flask
flask-cors
tensorflow
numpy
opencv-python
Pillow
gunicorn
